api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-alembic>=3.1.1",
    "flask-sqlalchemy-lite>=0.1.0",
]

[tool.uv.workspace]
members = [
    "temp",
]
